+++ src/constants/product.ts
@@ -1,0 +1,6 @@
+export const INITIAL_VALUES = {
+  name: "",
+  description: "",
+  price: 0,
+  tag: [],
+};
